+++ client/public/scripts/campaign/levels/debug1/level.js
@@ -1,3 +1,6 @@
 return (class extends LevelPlayerScript{
-    
+    async on_spawn_player(player){
+        //player.set_preset(await this.level.load_character({"path": "../../characters/maria.jsonc"}))
+        //player.set_preset(await this.level.load_character({"path": "../../characters/mark.jsonc"}))
+    }
 })