+++ common/scripts/definitions/items/guns.ts
@@ -85,19 +85,21 @@ export const bullets_factory={
 
             critical_mult: 1.25,
             obstacle_mult: 1,
-            falloff:0.85,
+            effective_range: 0.2,
+            falloff:0.8,
             tracer:tracers.medium
         }
     },
     sniper(power:number,tracer=tracers.large):BulletDef{
         return {
-            damage: 51 * power,
+            damage: 50 * power,
             range: 200 * (1 + (power - 1) * 0.5),
             speed: 55 * (1 + (power - 1) * 0.6),
 
             critical_mult: 1.1,
             obstacle_mult: 1.25,
-            falloff: 0.95,
+            falloff: 0.75,
+            effective_range:0.17,
             tracer: tracer,
         }
     },
@@ -110,6 +112,7 @@ export const bullets_factory={
             falloff:0.75,
             critical_mult:1.1,
             obstacle_mult:2,
+            effective_range:0.12,
             tracer:tracer
         }
     },
@@ -174,6 +177,7 @@ export const bullets_factory={
             damage:6*power,
             speed:30*(1 + (power - 1) * 0.4),
             range:55*(1 + (power - 1) * 0.2),
+            effective_range:0.17,
 
             falloff:0.85,
             critical_mult:1.2,
@@ -1344,11 +1348,12 @@ export function Guns_Default_Init():GunDef[]{
             },
             bullet:{
                 def:{
-                    damage:26,
+                    damage:25,
                     falloff:0.75,
                     range:165,
                     speed:50,
-                    tracer:tracers.large
+                    tracer:tracers.large,
+                    effective_range:0.17,
                 }
             },
             reload:{